update waypoint works, but saving waypoint doesn't add it to route

+++ WalkingToursApp/Waypoint.swift
@@ -15,7 +15,7 @@ class Waypoint: NSObject {
     var wpLat           :String?
     var wpLon           :String?
     var wpDescript      :String?
-    var wpStopNum       :Int8 = 0
+    var wpStopNum       :Int = 0
     var wpPicName       :String?
     var wpAudioName     :String?
     var objectId        :String?
